requirements.txt
asgiref==3.2.10
Django==3.1.1
django-ckeditor==6.0.0
django-cleanup==5.1.0
django-js-asset==1.2.2
django-widget-tweaks==1.4.8
Pillow==7.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pytz==2020.1
sqlparse==0.3.1
